// src/api/openapi/models/TripRecommendationRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TripRecommendationRouteRequestDto } from './TripRecommendationRouteRequestDto';
import {
    TripRecommendationRouteRequestDtoFromJSON,
    TripRecommendationRouteRequestDtoFromJSONTyped,
    TripRecommendationRouteRequestDtoToJSON,
    TripRecommendationRouteRequestDtoToJSONTyped,
} from './TripRecommendationRouteRequestDto';

/**
 * 
 * @export
 * @interface TripRecommendationRequestDto
 */
export interface TripRecommendationRequestDto {
    /**
     * 
     * @type {Array<TripRecommendationRouteRequestDto>}
     * @memberof TripRecommendationRequestDto
     */
    routes: Array<TripRecommendationRouteRequestDto>;
}

/**
 * Check if a given object implements the TripRecommendationRequestDto interface.
 */
export function instanceOfTripRecommendationRequestDto(value: object): value is TripRecommendationRequestDto {
    if (!('routes' in value) || value['routes'] === undefined) return false;
    return true;
}

export function TripRecommendationRequestDtoFromJSON(json: any): TripRecommendationRequestDto {
    return TripRecommendationRequestDtoFromJSONTyped(json, false);
}

export function TripRecommendationRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TripRecommendationRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'routes': ((json['routes'] as Array<any>).map(TripRecommendationRouteRequestDtoFromJSON)),
    };
}

export function TripRecommendationRequestDtoToJSON(json: any): TripRecommendationRequestDto {
    return TripRecommendationRequestDtoToJSONTyped(json, false);
}

export function TripRecommendationRequestDtoToJSONTyped(value?: TripRecommendationRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'routes': ((value['routes'] as Array<any>).map(TripRecommendationRouteRequestDtoToJSON)),
    };
}