// src/api/openapi/models/TripRecommendationRouteRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TripRecommendationRouteRequestDto
 */
export interface TripRecommendationRouteRequestDto {
    /**
     * 
     * @type {number}
     * @memberof TripRecommendationRouteRequestDto
     */
    directionTimeSeconds: number;
    /**
     * 
     * @type {number}
     * @memberof TripRecommendationRouteRequestDto
     */
    previousPauseTimeSeconds: number;
}

/**
 * Check if a given object implements the TripRecommendationRouteRequestDto interface.
 */
export function instanceOfTripRecommendationRouteRequestDto(value: object): value is TripRecommendationRouteRequestDto {
    if (!('directionTimeSeconds' in value) || value['directionTimeSeconds'] === undefined) return false;
    if (!('previousPauseTimeSeconds' in value) || value['previousPauseTimeSeconds'] === undefined) return false;
    return true;
}

export function TripRecommendationRouteRequestDtoFromJSON(json: any): TripRecommendationRouteRequestDto {
    return TripRecommendationRouteRequestDtoFromJSONTyped(json, false);
}

export function TripRecommendationRouteRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TripRecommendationRouteRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'directionTimeSeconds': json['directionTimeSeconds'],
        'previousPauseTimeSeconds': json['previousPauseTimeSeconds'],
    };
}

export function TripRecommendationRouteRequestDtoToJSON(json: any): TripRecommendationRouteRequestDto {
    return TripRecommendationRouteRequestDtoToJSONTyped(json, false);
}

export function TripRecommendationRouteRequestDtoToJSONTyped(value?: TripRecommendationRouteRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'directionTimeSeconds': value['directionTimeSeconds'],
        'previousPauseTimeSeconds': value['previousPauseTimeSeconds'],
    };
}