// src/api/openapi/models/TripRecommendationResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TripRecommendationType } from './TripRecommendationType';
import {
    TripRecommendationTypeFromJSON,
    TripRecommendationTypeFromJSONTyped,
    TripRecommendationTypeToJSON,
    TripRecommendationTypeToJSONTyped,
} from './TripRecommendationType';
import type { TripRecommendationRouteResponseDto } from './TripRecommendationRouteResponseDto';
import {
    TripRecommendationRouteResponseDtoFromJSON,
    TripRecommendationRouteResponseDtoFromJSONTyped,
    TripRecommendationRouteResponseDtoToJSON,
    TripRecommendationRouteResponseDtoToJSONTyped,
} from './TripRecommendationRouteResponseDto';

/**
 * 
 * @export
 * @interface TripRecommendationResponseDto
 */
export interface TripRecommendationResponseDto {
    /**
     * 
     * @type {TripRecommendationType}
     * @memberof TripRecommendationResponseDto
     */
    type: TripRecommendationType;
    /**
     * 
     * @type {Array<TripRecommendationRouteResponseDto>}
     * @memberof TripRecommendationResponseDto
     */
    routes: Array<TripRecommendationRouteResponseDto>;
    /**
     * 
     * @type {number}
     * @memberof TripRecommendationResponseDto
     */
    reduceRoutesHours: number;
    /**
     * 
     * @type {number}
     * @memberof TripRecommendationResponseDto
     */
    reduceTimeHours: number;
}



/**
 * Check if a given object implements the TripRecommendationResponseDto interface.
 */
export function instanceOfTripRecommendationResponseDto(value: object): value is TripRecommendationResponseDto {
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('routes' in value) || value['routes'] === undefined) return false;
    if (!('reduceRoutesHours' in value) || value['reduceRoutesHours'] === undefined) return false;
    if (!('reduceTimeHours' in value) || value['reduceTimeHours'] === undefined) return false;
    return true;
}

export function TripRecommendationResponseDtoFromJSON(json: any): TripRecommendationResponseDto {
    return TripRecommendationResponseDtoFromJSONTyped(json, false);
}

export function TripRecommendationResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TripRecommendationResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'type': TripRecommendationTypeFromJSON(json['type']),
        'routes': ((json['routes'] as Array<any>).map(TripRecommendationRouteResponseDtoFromJSON)),
        'reduceRoutesHours': json['reduce_routes_hours'],
        'reduceTimeHours': json['reduce_time_hours'],
    };
}

export function TripRecommendationResponseDtoToJSON(json: any): TripRecommendationResponseDto {
    return TripRecommendationResponseDtoToJSONTyped(json, false);
}

export function TripRecommendationResponseDtoToJSONTyped(value?: TripRecommendationResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': TripRecommendationTypeToJSON(value['type']),
        'routes': ((value['routes'] as Array<any>).map(TripRecommendationRouteResponseDtoToJSON)),
        'reduce_routes_hours': value['reduceRoutesHours'],
        'reduce_time_hours': value['reduceTimeHours'],
    };
}